+++ Sources/HAP/Server/Device.swift
@@ -76,7 +76,12 @@ public class Device {
         // characteristics within an HAP accessory object. The instance ID for
         // each object must be unique for the lifetime of the server/ client
         // pairing.
-        var idGenerator = (1...UInt64.max).makeIterator()
+
+        // TODO: SwiftFoundation in Swift 4.0 cannot encode/decode UInt64,
+        // which is the data-type we wanted to use here. We can change it back
+        // to UInt64 once the following commit has made it into a release:
+        // https://github.com/apple/swift-corelibs-foundation/commit/64b67c91479390776c43a96bd31e4e85f106d5e1
+        var idGenerator = (1...Int.max).makeIterator()
         for accessory in accessories {
             accessory.device = self
             accessory.aid = idGenerator.next()!
